move common application and agent to pkg/metaserver
